requirements.txt
chromadb>=0.5.5
sentence-transformers>=3.0.1
openai>=1.30.0
requests>=2.31.0
pydantic>=2.7.0
tqdm>=4.66.4
beautifulsoup4>=4.12.3
feedparser>=6.0.11
gradio>=4.44.0
pandas>=2.2.2
python-dotenv>=1.0.1
twilio>=9.0.0
typing-extensions>=4.9.0
